UID and GID must be uint32

+++ cloud/disk_manager/internal/pkg/facade/filesystem_snapshot_transfer_test/filesystem_snapshot_transfer_test.go
@@ -79,8 +79,8 @@ type comparableNode struct {
 	Mode  uint32
 	Type  uint32
 	Links uint32
-	UID   uint64
-	GID   uint64
+	UID   uint32
+	GID   uint32
 }
 
 func listComparableNodes(
